+++ Anubis.LC.ExtraDays/Helpers/LethalConfigHelper.cs
@@ -6,16 +6,18 @@ using UnityEngine;
 using LethalConfig;
 using System.Collections.Generic;
 using System;
+using LethalConfig.ConfigItems.Options;
+using Anubis.LC.ExtraDays.ModNetwork;
 
 namespace Anubis.LC.ExtraDays.Helpers
 {
     public static class LethalConfigHelper
     {
-        public static Dictionary<string, Dictionary<string, ConfigEntry<bool>>> SaveFilesConfigurations;
+        public static Dictionary<string, Dictionary<string, ConfigEntryBase>> SaveFilesConfigurations;
 
         public static void SetLehalConfig(ConfigFile config)
         {
-            SaveFilesConfigurations = new Dictionary<string, Dictionary<string, ConfigEntry<bool>>>();
+            SaveFilesConfigurations = new Dictionary<string, Dictionary<string, ConfigEntryBase>>();
 
             int numOfSaveFiles = 0;
             foreach (string file in ES3.GetFiles())
@@ -35,13 +37,28 @@ namespace Anubis.LC.ExtraDays.Helpers
                     if (ES3.FileExists(file) && file.StartsWith("LCSaveFile"))
                     {
                         var currentIndex = i++;
-                        var configurationForSaveFile = new Dictionary<string, ConfigEntry<bool>>();
+                        var configurationForSaveFile = new Dictionary<string, ConfigEntryBase>();
                         var correlatedPrice = config.Bind($"Save File {currentIndex}", "Use price correlated calculation?", true, "This determines if the price to buy an extra day will be constant value (350 credits) or correlated to the quota (dynamic)");
                         var buyingRate = config.Bind($"Save File {currentIndex}", "Use reduce buying rate?", false, "This determines if the buying rate will reduce a bit after buying an extra day");
+                        var extraDayPrice = config.Bind($"Save File {currentIndex}", "Setup extra day price", 350, "This configure the price of an extra day only if `correlated calculation` field is OFF");
+                        extraDayPrice.SettingChanged += (obj, args) =>
+                        {
+                            if (Networking.Instance && RoundManager.Instance.NetworkManager.IsHost)
+                            {
+                                Networking.Instance.SetExtraDaysPriceServerRpc(extraDayPrice.Value);
+                            }
+                        };
                         LethalConfigManager.AddConfigItem(new BoolCheckBoxConfigItem(correlatedPrice, false));
+                        LethalConfigManager.AddConfigItem(new IntSliderConfigItem(extraDayPrice, new IntSliderOptions()
+                        {
+                            Min = 250,
+                            Max = 1700,
+                            RequiresRestart = false,
+                        }));
                         LethalConfigManager.AddConfigItem(new BoolCheckBoxConfigItem(buyingRate, false));
                         configurationForSaveFile.Add("correlatedPrice", correlatedPrice);
                         configurationForSaveFile.Add("buyingRate", buyingRate);
+                        configurationForSaveFile.Add("extraDayPrice", extraDayPrice);
                         SaveFilesConfigurations.Add(file, configurationForSaveFile);
                     }
                 }
@@ -51,13 +68,28 @@ namespace Anubis.LC.ExtraDays.Helpers
                 for (int i = 0; i < 3; i++)
                 {
                     var currentIndex = i + 1;
-                    var configurationForSaveFile = new Dictionary<string, ConfigEntry<bool>>();
+                    var configurationForSaveFile = new Dictionary<string, ConfigEntryBase>();
                     var correlatedPrice = config.Bind($"Save File {currentIndex}", "Use price correlated calculation?", true, "This determines if the price to buy an extra day will be constant value (350 credits) or correlated to the quota (dynamic)");
                     var buyingRate = config.Bind($"Save File {currentIndex}", "Use reduce buying rate?", false, "This determines if the buying rate will reduce a bit after buying an extra day");
+                    var extraDayPrice = config.Bind($"Save File {currentIndex}", "Setup extra day price", 350, "This configure the price of an extra day only if `correlated calculation` field is OFF");
+                    extraDayPrice.SettingChanged += (obj, args) =>
+                    {
+                        if (Networking.Instance && RoundManager.Instance.NetworkManager.IsHost)
+                        {
+                            Networking.Instance.SetExtraDaysPriceServerRpc(extraDayPrice.Value);
+                        }
+                    };
                     LethalConfigManager.AddConfigItem(new BoolCheckBoxConfigItem(correlatedPrice, false));
+                    LethalConfigManager.AddConfigItem(new IntSliderConfigItem(extraDayPrice, new IntSliderOptions()
+                    {
+                        Min = 250,
+                        Max = 1700,
+                        RequiresRestart = false,
+                    }));
                     LethalConfigManager.AddConfigItem(new BoolCheckBoxConfigItem(buyingRate, false));
                     configurationForSaveFile.Add("correlatedPrice", correlatedPrice);
                     configurationForSaveFile.Add("buyingRate", buyingRate);
+                    configurationForSaveFile.Add("extraDayPrice", extraDayPrice);
                     SaveFilesConfigurations.Add($"LCSaveFile{i}", configurationForSaveFile);
                 }
             }
@@ -66,19 +98,27 @@ namespace Anubis.LC.ExtraDays.Helpers
             LethalConfigManager.SetModDescription("Allows the player to purchase an extra day via the terminal. This mod's uniqueness is that it tries to be more realistic with the game, as it is not trying to modify the main logic of the deadline. It tries to add functionality that will give you extends for the deadline but with some price on it.");
         }
 
-        public static Dictionary<string, ConfigEntry<bool>> GetConfigForSaveFile()
+        public static Dictionary<string, ConfigEntryBase> GetConfigForSaveFile()
         {
             SaveFilesConfigurations.TryGetValue(GameNetworkManager.Instance.currentSaveFileName, out var config);
             if (config == null)
             {
                 ModStaticHelper.Logger.LogError("Could not find save file. Using defaults configuration for save files");
-                var configurationForSaveFile = new Dictionary<string, ConfigEntry<bool>>();
+                var configurationForSaveFile = new Dictionary<string, ConfigEntryBase>();
                 var correlatedPrice = ExtraDaysToDeadlinePlugin.Instance.Config.Bind($"Save File 1", "Use price correlated calculation?", true, "This determines if the price to buy an extra day will be constant value (350 credits) or correlated to the quota (dynamic)");
                 var buyingRate = ExtraDaysToDeadlinePlugin.Instance.Config.Bind($"Save File 1", "Use reduce buying rate?", false, "This determines if the buying rate will reduce a bit after buying an extra day");
+                var extraDayPrice = ExtraDaysToDeadlinePlugin.Instance.Config.Bind($"Save File 1", "Setup extra day price", 350, "This configure the price of an extra day only if `correlated calculation` field is OFF");
                 LethalConfigManager.AddConfigItem(new BoolCheckBoxConfigItem(correlatedPrice, false));
+                LethalConfigManager.AddConfigItem(new IntSliderConfigItem(extraDayPrice, new IntSliderOptions()
+                {
+                    Min = 250,
+                    Max = 1700,
+                    RequiresRestart = false,
+                }));
                 LethalConfigManager.AddConfigItem(new BoolCheckBoxConfigItem(buyingRate, false));
                 configurationForSaveFile.Add("correlatedPrice", correlatedPrice);
                 configurationForSaveFile.Add("buyingRate", buyingRate);
+                configurationForSaveFile.Add("extraDayPrice", extraDayPrice);
                 return configurationForSaveFile;
             }
             return config;